Modularity
This commit restructures the project significantly

We now support:
- TypeScript exclusive source
- A build that supports CommonJS and ESM
- A build that exports types
- Both flat "top level" exports at @sero.run/sero and submodule specific exports at ex. @sero.run/sero/cds-hooks/service

Importantly, the examples folder can no longer be run from the root - anything in the examples file has to be isolated completely and have its own dependencies (start doesnt exist as a command anymore)
