--- jest.config.ts
@@ -1,9 +1,0 @@
-import type {Config} from '@jest/types';
-
-const config: Config.InitialOptions = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  rootDir: 'src'
-};
-
-export default config;
