+++ src/cds-hooks/index.ts
@@ -1,24 +1,9 @@
 /**
- * @module cds-hooks
+ * @module @sero.run/sero/cds-hooks
  */
 
-import { FastifyInstance } from "fastify";
-import { Config } from "..";
-import routes from "./routes";
-import Service from "./service";
-
-export default (config: Config, http: FastifyInstance): void => {
-  routes(http, config.cdsHooks);
-}
-
-export function getService(services: Service[], id: string): Service | undefined {
-  return services.find((service) => service.id == id)
-}
-
-export type Hooks = "patient-view" | "order-sign" | "order-select" | "order-review" | "medication-prescribe" | "encounter-start" | "encounter-discharge" | "appointment-book"
-
-export { default as Service } from "./service";
+export { default as mount } from "./routes";
+export { default as Service, HookRequest, HookResponse } from "./service";
 export { default as Card } from "./card";
 export { default as Suggestion } from "./suggestion";
-
-export class NoDecisionResponse extends Error { }
+export { NoDecisionResponse } from "./util";
