+++ src/cds-hooks/suggestion.ts
@@ -1,6 +1,7 @@
 import { randomUUID } from "crypto";
+import { SystemAction } from "./util";
 
-export default class Suggestion implements CDSHooks.Suggestion {
+export default class Suggestion implements Suggestion {
   /**
    * Unique identifier of the card. MAY be used for auditing and logging cards
    * and SHALL be included in any subsequent calls to the CDS service's feedback
@@ -25,12 +26,20 @@ export default class Suggestion implements CDSHooks.Suggestion {
    * all actions are logically AND'd together, such that a user selecting a
    * suggestion selects all of the actions within it.
    */
-  actions?: CDSHooks.SystemAction[];
+  actions?: SystemAction[];
 
-  constructor(options: Partial<CDSHooks.Suggestion> & { label: string }) {
+  constructor(options: Partial<Suggestion> & { label: string }) {
     this.uuid = options.uuid || randomUUID();
     this.label = options.label;
     this.isRecommended = options.isRecommended;
     this.actions = options.actions;
   }
+}
+
+export interface AcceptedSuggestion {
+  /**
+   * The card.suggestion.uuid from the CDS Hooks response. Uniquely identifies
+   * the suggestion that was accepted.
+   */
+  id: string
 }
