+++ src/cds-hooks/util.ts
@@ -1,6 +1,7 @@
 import { Service } from "./index";
-import Ajv, { ValidateFunction, ErrorObject } from "ajv";
-import { Hooks } from "./index"
+import Ajv from "ajv";
+import { ValidateFunction, ErrorObject } from "ajv";
+import { HookRequest } from "./service";
 
 /**
  * Validates that the hook request dynmically (depending on the hook and service at runtime):
@@ -11,7 +12,7 @@ import { Hooks } from "./index"
  * @param reply
  * @returns Error | undefined
  */
-export function validateHookRequest(hookRequest: CDSHooks.HookRequest<Record<string, unknown>>, service: Service): Error | undefined {
+export function validateHookRequest(hookRequest: HookRequest<Record<string, unknown>>, service: Service): Error | undefined {
   // @ts-expect-error hookRequest.hook needs to be co-erced into the type...
   const context = validateContext(hookRequest, hookRequest.hook);
   const prefetch = validatePrefetch(hookRequest, service);
@@ -63,7 +64,7 @@ function schemaErrorFormatter(errors: ErrorObject[], dataVar: string) {
  * @param service 
  * @returns 
  */
-function validatePrefetch(request: CDSHooks.HookRequest<Record<string, any>>, service: Service): ValidateFunction {
+function validatePrefetch(request: HookRequest<Record<string, any>>, service: Service): ValidateFunction {
   const ajv = new Ajv({
     removeAdditional: false,
     useDefaults: true,
@@ -117,7 +118,7 @@ function validatePrefetch(request: CDSHooks.HookRequest<Record<string, any>>, se
   return compiled
 }
 
-function validateContext(request: CDSHooks.HookRequest<Record<string, any>>, hook: Hooks): ValidateFunction {
+function validateContext(request: HookRequest<Record<string, any>>, hook: Hooks): ValidateFunction {
   const ajv = new Ajv({
     removeAdditional: true,
     useDefaults: true,
@@ -230,4 +231,66 @@ function validateContext(request: CDSHooks.HookRequest<Record<string, any>>, hoo
   const compiled = ajv.compile(schema)
   compiled(context)
   return compiled
+}
+
+export function getService(services: Service[], id: string): Service | undefined {
+  return services.find((service) => service.id == id)
+}
+
+export class NoDecisionResponse extends Error {}
+
+export interface FhirAuthorization {
+  /**
+   * This is the OAuth 2.0 access token that provides access to the FHIR
+   * server.
+   */
+  access_token: string
+
+  token_type: 'Bearer'
+
+  /**
+   * The lifetime in seconds of the access token.
+   */
+  expires_in: number
+
+  /**
+   * The scopes the access token grants the CDS Service.
+   */
+  scope: string
+
+  /**
+   * The OAuth 2.0 client identifier of the CDS Service, as registered with the
+   * CDS Client's authorization server.
+   */
+  subject: string
+}
+
+export type Hooks = "patient-view" | "order-sign" | "order-select" | "order-review" | "medication-prescribe" | "encounter-start" | "encounter-discharge" | "appointment-book"
+
+export interface SystemAction {
+  /**
+   * The type of action being performed. Allowed values are: create, update, delete.
+   */
+  type: 'create' | 'update' | 'delete'
+
+  /**
+   * Human-readable description of the suggested action MAY be presented to the
+   * end-user.
+   */
+  description: string
+
+  /**
+   * A FHIR resource. When the type attribute is create, the resource attribute
+   * SHALL contain a new FHIR resource to be created. For update, this holds
+   * the updated resource in its entirety and not just the changed fields. Use
+   * of this field to communicate a string of a FHIR id for delete suggestions
+   * is DEPRECATED and resourceId SHOULD be used instead.
+   */
+  resource?: any
+
+  /**
+   * A relative reference to the relevant resource. SHOULD be provided when the
+   * type attribute is delete.
+   */
+  resourceId?: string
 }
