+++ src/index.ts
@@ -1,8 +1,26 @@
 /**
- * @module sero/sero
+ * @module @sero.run/sero
  */
 
-export { default as Rest } from "./rest";
-export { default as CDSHooks } from "./cds-hooks";
-export { default as Config } from "./config";
-export { default as Http } from "./http";
+export {
+  mount as Rest
+} from "./rest";
+
+export {
+  mount as CDSHooks,
+  HookRequest,
+  HookResponse,
+  Card,
+  Service,
+  Suggestion,
+  NoDecisionResponse,
+} from "./cds-hooks";
+
+export {
+  default as Config
+} from "./config";
+
+export {
+  default as Http,
+  start
+} from "./http";
