--- src/types/index.d.ts
@@ -1,2 +1,0 @@
-import "./cds-hooks/index.d.ts"
-import "./smart/index.d.ts"
