+++ tsconfig/test.json
@@ -1,0 +1,6 @@
+{
+  "extends": "./base.json",
+  "compilerOptions": {
+    "esModuleInterop": true
+  }
+}
